Set signed update file name to match factory file name
When generating signed updates we set the file name to match the factory
name so it's easier to be clear about which update you're applying.

We also generate a hash of the file for validation, especially useful if
we're moving the update around and want to validate it's integrity before
using the new update image.