Fix update process for angler
The previous update-binary was not successfully getting the properties,
and therefore wasn't detecting that the device was an angler device.

This binary has been taken from here:

    https://legacy.copperhead.co/angler-ota_update-2017.09.29.01.zip

The end of the updater-script is causing some odd output when applied,
so following the example from the stock one the last command is a
set_progress(), and the bang (!) is taken out of the ui_print() in case
that was causing any issues too.